django/requirements.txt
Django
psycopg2
gunicorn
pyopenssl
pandas
xlrd
requests
django-elasticsearch-dsl
djangorestframework
pyexcel
pyexcel-io
